requirements.txt
yfinance
pandas
numpy
python-docx
groq
